moved logic tests to a folder
